Remove filename restriction, Add audio workaround
The audio sync may be slightly off, but this allows there to be sound
to accompany the video.
The file is still loaded from the same directory, but no longer has to
be named hologram. The current setup will only play the first q3d file
found in a directory.